#9 rename modules and separate exemple and src
