+++ examples/cases/ccm_example_no_sub_cluster.py
