+++ examples/cases/illustration_1.py
