+++ examples/cases/illustration_2.py
