+++ examples/cases/small.py
