+++ examples/cases/synthetic_blocks.py
