+++ examples/configuration.py
@@ -33,35 +33,6 @@
 # Questions? Contact darma@sandia.gov
 #
 
-class Config:
-    """Config object"""
-    def __init__(
-        self,
-        is_fmwp: bool,
-        alpha: float,
-        beta: float,
-        gamma: float,
-        delta: float,
-        bounded_memory:bool,
-        preserve_clusters:bool
-    ):
-        print(f"\n# Initializing {'FMWP' if is_fmwp else 'COMCP'} configuration with:")
-        print(f"  alpha = {alpha}")
-        print(f"  beta = {beta}")
-        print(f"  gamma = {gamma}")
-        print(f"  delta = {delta}")
-        print(f"  with{'' if bounded_memory else 'out'} rank memory upper bound")
-        if preserve_clusters:
-            print("  while preserving block clusters")
-        self.is_fmwp = is_fmwp
-        self.is_comcp = not is_fmwp
-        self.alpha = alpha
-        self.beta = beta
-        self.gamma = gamma
-        self.delta = delta
-        self.bounded_memory = bounded_memory
-        self.preserve_clusters = preserve_clusters
-
 class ExampleConfig:
     """Examples Configs"""
     def __init__(
@@ -158,20 +129,7 @@ class Examples:
                 classname = 'Illustration1'
             ),
             ExampleConfig(
-                filename = 'Illustration2',
-                classname = '5'
+                filename = 'illustration_2',
+                classname = 'Illustration2'
             )
         ]
-
-class Parameters:
-    """Parameters"""
-
-    @staticmethod
-    def defaults():
-        """Default parameters list"""
-        return {
-            "alpha": 1.0,
-            "beta": 0.0,
-            "gamma": 0.0,
-            "delta": 0.0
-        }
